requirements.txt
asgiref
certifi
charset-normalizer
crispy-bootstrap5
dj-database-url
Django
django-cors-headers
django-crispy-forms
django-email
django-environ
django-filter
djangorestframework
djangorestframework-simplejwt
idna
Markdown
pillow
psycopg2
psycopg2-binary
PyJWT
pytz
requests
rest-framework-simplejwt
sqlparse
typing_extensions
tzdata
urllib3